Add Hooggaa experiment
Also reorganize so that it's easier to add experiments in the future.
The basic structure now is that each project has a top level html page,
and it's resources are appropriately named in the corresponding css,
scripts, and assets folders.

This structure allows the paths to work in the `pub serve` case, as well
as the appengine static rewrite server case.