Basic functionality working
Still need to clean up the dart code a bit, as well as figure out how to
get the animation of the flip cards to work properly. The original
example (https://codepen.io/doriancami/pen/jEJvaV) used the TweenMax
library, but I'm trying to recreate the effect via native dart.