Add smartcam project
This is meant as a starting point for working on a custom ML model to
detect various features in streaming webcam video and take actions based
off of them.

This version does simple person detection, and can trigger video
recording based on whether a person is present in the frame or not.