Move notebook contents to seperate folder
The idea is to add support for a url parameter so that I can load/ share
saved (and uploaded) work. If I want to reference data, though, I need
to either upload it along with the rest of the CDN code, or figure out a
CORS compatible solution that lets me store it elsewhere.