Fix time dilation bug affecting longer pomodoro sessions
Basically, some browsers automatically throttle background/ non-active
tabs, so if you are using the pomodoro timer in a background tab (or
progressive web app that doesn't have focus), the old code would start
to lag behind the actually elapsed clock time, since it was simply
counting down from the originally input duration, and while throttled,
the count went slower/ stopped.

To fix that "issue", I added a `trueElapsedSeconds` function which
computes the elapsed time based on a timestamp from when the pomodoro
timer was started. If the "countdown" elapsed time differs from the true
elapsed time by more than some small delta, the new code will
automatically update the countdown elapsed time to its true value.

I put "issue" in quotes above because for my own usage this has actually
turned out to be a feature, wherein my pomodoros stretched for much
longer than the originally allocated 30 minutes, and I was able to get a
larger block of productive work in. As such, I plan to keep the "time
dilation" effect as the default behavior, but added a URL param flag
(`t=1`) to turn on "true time" calculation for the pomodoro session.