Add custom RSS reader (gazette) frontend.
This change adds a frontend view to replicate the experience of reading a
newspaper, such that I can read articles in a more "traditional" medium
and hopefully clear out my email inbox backlog in the New Year.

I have included a set of hardcoded sample of articles to get a sense of the
layout on my various device screens, and functionality to clear those
samples when a proper "gazette payload" is present.